Functions

KLU.KLUFactorization — Type

KLUFactorization <: Factorization

Matrix factorization type of the KLU factorization of a sparse matrix A. This is the return type of klu, the corresponding matrix factorization function.

The factors can be obtained from K::KLUFactorization via K.L, K.U and K.F See the klu docs for more information.

You typically should not construct this directly, instead use klu.

KLU.klu_common — Type

Data structure for parameters of and statistics generated by KLU functions.

This is the Int32 version of klu_l_common.

KLU.klu_l_common — Type

Data structure for parameters of and statistics generated by KLU functions.

Fields

tol::Float64: Partial pivoting tolerance for diagonal preference

btf::Int64: If btf != 0 use BTF pre-ordering

ordering::Int64: If ordering == 0 use AMD to permute, if ordering == 1 use COLAMD,

if ordering == 3 use the user provided ordering function.

scale::Int64: If scale == 1 then A[:,i] ./= sum(abs.(A[:,i])), if scale == 2 then

A[:,i] ./= maximum(abs.(A[:,i])). If scale == 0 no scaling is done, and the input is checked for errors if scale >= 0.

See the KLU User Guide for more information.

KLU.condest — Method

condest(K::KLUFactorization)

Accurately estimate the 1-norm condition number of the factorization.

KLU.condest — Method

condest(K::KLUFactorization)

Accurately estimate the 1-norm condition number of the factorization.

KLU.condest — Method

condest(K::KLUFactorization)

Accurately estimate the 1-norm condition number of the factorization.

KLU.condest — Method

condest(K::KLUFactorization)

Accurately estimate the 1-norm condition number of the factorization.

KLU.klu — Method

klu(A::SparseMatrixCSC) -> K::KLUFactorization
klu(n, colptr::Vector{Ti}, rowval::Vector{Ti}, nzval::Vector{Tv}) -> K::KLUFactorization

Compute the LU factorization of a sparse matrix A using KLU[ACM907].

For sparse A with real or complex element type, the return type of K is KLUFactorization{Tv, Ti}, with Tv = Float64 or ComplexF64 respectively and Ti is an integer type (Int32 or Int64).

The individual components of the factorization K can be accessed by indexing:

Component	Description

L	L (lower triangular) part of LU of the diagonal blocks

U	U (upper triangular) part of LU of the diagonal blocks

F	F (upper triangular) part of LU + F, the off-diagonal blocks

p	right permutation Vector

q	left permutation Vector

Rs	Vector of scaling factors

The relation between K and A is

K.L * K.U + K.F == K.Rs \ A[K.p, K.q]

K further supports the following functions:

LinearAlgebra.\

Note

klu(A::SparseMatrixCSC) uses the KLU[ACM907] library that is part of SuiteSparse. As this library only supports sparse matrices with Float64 or ComplexF64 elements, lu converts A into a copy that is of type SparseMatrixCSC{Float64} or SparseMatrixCSC{ComplexF64} as appropriate.

KLU.klu_factor! — Function

klu_factor!(K::KLUFactorization)

Factor K into components K.L, K.U, and K.F. This function will perform both the symbolic and numeric steps of factoriation on an existing KLUFactorization instance.

The K.common struct can be used to modify certain options and parameters, see the KLU documentation or klu_common for more information.

KLU.rcond — Method

rcond(K::KLUFactorization)

Cheaply estimate the reciprocal condition number.

KLU.rcond — Method

rcond(K::KLUFactorization)

Cheaply estimate the reciprocal condition number.

KLU.rcond — Method

rcond(K::KLUFactorization)

Cheaply estimate the reciprocal condition number.

KLU.rcond — Method

rcond(K::KLUFactorization)

Cheaply estimate the reciprocal condition number.

KLU.rgrowth — Method

rgrowth(K::KLUFactorization)

Calculate the reciprocal pivot growth.

KLU.rgrowth — Method

rgrowth(K::KLUFactorization)

Calculate the reciprocal pivot growth.

KLU.rgrowth — Method

rgrowth(K::KLUFactorization)

Calculate the reciprocal pivot growth.

KLU.rgrowth — Method

rgrowth(K::KLUFactorization)

Calculate the reciprocal pivot growth.

ACM907Davis, Timothy A., & Palamadai Natarajan, E. (2010). Algorithm 907: KLU, A Direct Sparse Solver for Circuit Simulation Problems. ACM Trans. Math. Softw., 37(3). doi:10.1145/1824801.1824814