KLU.jl

This package wraps the SuiteSparse KLU solver. A description from the publication:

KLU is a set of routines for solving sparse linear systems of equations. It is particularly well-suited to matrices arising in SPICE-like circuit simulation applications. It relies on a permutation to block triangular form (BTF), several methods for finding a fill-reducing ordering (variants of approximate minimum degree, and nested dissection), and a sparse left-looking LU factorization method to factorize each block.

Installation

using Pkg
Pkg.add("KLU")

Usage

julia> using KLU
julia> using SparseArrays
julia> A = sprand(15, 15, 0.45);
julia> factor = klu(A);
julia> b = rand(15);
julia> x = factor \ b
15-element Vector{Float64}:
  22.501711574339556
  27.80357152530039
 -37.08488119597194
  61.470604418791346
  -0.49109038415655193
  38.03163985997178
   1.2417021657549396
 -51.154984106483155
   0.9488554483842341
   1.587572050319512
 -31.251127382496147
   1.4508846663863602
 -25.866070230462796
  -5.46661465074706
  10.688731342850245

Citing this package

Please cite both SuiteSparse as well as KLU:

Exported API

KLU.klu — Function

klu(A::SparseMatrixCSC) -> K::KLUFactorization
klu(n, colptr::Vector{Ti}, rowval::Vector{Ti}, nzval::Vector{Tv}) -> K::KLUFactorization

Compute the LU factorization of a sparse matrix A using KLU[ACM907].

For sparse A with real or complex element type, the return type of K is KLUFactorization{Tv, Ti}, with Tv = Float64 or ComplexF64 respectively and Ti is an integer type (Int32 or Int64).

The individual components of the factorization K can be accessed by indexing:

Component	Description

L	L (lower triangular) part of LU of the diagonal blocks

U	U (upper triangular) part of LU of the diagonal blocks

F	F (upper triangular) part of LU + F, the off-diagonal blocks

p	right permutation Vector

q	left permutation Vector

Rs	Vector of scaling factors

The relation between K and A is

K.L * K.U + K.F == K.Rs \ A[K.p, K.q]

K further supports the following functions:

LinearAlgebra.\

Arguments

A::SparseMatrixCSC or n::Integer, colptr::Vector{Ti}, rowval::Vector{Ti}, nzval::Vector{Tv}: The sparse matrix or the zero-based sparse matrix components to be factored.

check::Bool: If true (default) check for errors after the factorization. If false errors must be checked by the user with klu.common.status.

allowsingular::Bool: If true (default false) allow the factorization to proceed even if the matrix is singular. Note that this will allow for

silent divide by zero errors in subsequent solve! or ldiv! calls if singularity is not checked by the user with klu.common.status == KLU.KLU_SINGULAR

Note

klu(A::SparseMatrixCSC) uses the KLU[ACM907] library that is part of SuiteSparse. As this library only supports sparse matrices with Float64 or ComplexF64 elements, lu converts A into a copy that is of type SparseMatrixCSC{Float64} or SparseMatrixCSC{ComplexF64} as appropriate.

ACM907Davis, Timothy A., & Palamadai Natarajan, E. (2010). Algorithm 907: KLU, A Direct Sparse Solver for Circuit Simulation Problems. ACM Trans. Math. Softw., 37(3). doi:10.1145/1824801.1824814